requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
magic-filter==1.0.12
pypdf==4.1.0
aiogram_i18n==1.3.4
Babel==2.13.1
pycurl==7.45.3
pathlib==1.0.1
python-dotenv==1.0.1